don't re-encode links to response's encoding
lxml may extract a link of unicode type
taking a better guess on its encoding.
If an extracted link contains symbols unsupported by response_encoding
encoding the link to it will fail.

Taking advantage of:

- lxml's better guess on the encoding
- RFC3986's requirement for ascii-only urls (links are an other thing)
- urljoin's unicode compatibility
- the fact that unicode links get canonicalized anyway

I convert the link only to unicode in python 2 and 3
decoding with the response's encoding when needed.